Protocols record if they succeeded
We need to know if a protocol's result corresponds to the experimental
reference or the reference from AA simulations. Now, the protocol
records if the results matches well enough the reference in the
SUCCESS_CODE file as an integer. This integer works like a POSIX exit
code: it is 0 for a success, and any non-zero value for a failure.